CR: Reorganise files
